# docs/swagger.yaml
definitions:
  models.CreateResourceResponse:
    properties:
      message:
        type: string
    type: object
  models.Resource:
    properties:
      path:
        type: string
      tags:
        items:
          type: string
        type: array
      title:
        type: string
      type:
        type: string
    required:
    - path
    - tags
    - title
    - type
    type: object
  models.SearchEngineRequest:
    properties:
      keyword:
        type: string
    required:
    - keyword
    type: object
  models.SearchEngineResult:
    properties:
      id:
        type: string
      title:
        type: string
    type: object
info:
  contact: {}
paths:
  /createNewResource:
    post:
      consumes:
      - multipart/form-data
      description: This route uses to create a new resource it accepts form data with
        your file form data contains Title as string, Type as string (only supprt
        this values["HTML Pages", "image"]), and Tags as array of string
      parameters:
      - description: File
        in: formData
        name: file
        required: true
        type: file
      - in: formData
        items:
          type: string
        name: tags
        required: true
        type: array
      - in: formData
        name: title
        required: true
        type: string
      - in: formData
        name: type
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.CreateResourceResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.CreateResourceResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.CreateResourceResponse'
      summary: Create a new resource
  /resource/{id}:
    get:
      consumes:
      - application/json
      description: This route uses to Get resource information in order to visualise
        the resource
      parameters:
      - description: resource id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Resource'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.CreateResourceResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.CreateResourceResponse'
      summary: Get resource information
  /search:
    post:
      consumes:
      - application/json
      description: This route uses to Search for a keyword in a tags
      parameters:
      - description: Search for a resource
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.SearchEngineRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.SearchEngineResult'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.CreateResourceResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.CreateResourceResponse'
      summary: Search for a keyword
swagger: "2.0"